perf(dlrmamba): parallel block-scan SS2D + bf16 SVD fix
- Replace sequential scan with block-wise parallel scan (64-step blocks)
- Conv1d cross-block mixing for inter-block dependencies
- Fix SVD for bf16 (cast to float32 for torch.linalg.svd)
- Set bs=24, image_size=320 for 69% VRAM on L4

Built with ANIMA by Robot Flow Labs

+++ src/anima_dlrmamba/losses.py
@@ -21,7 +21,7 @@ class StructureAwareDistillationLoss(nn.Module):
 
     @staticmethod
     def svd_alignment_loss(u_s: torch.Tensor, v_s: torch.Tensor, a_t: torch.Tensor) -> torch.Tensor:
-        u_t, s_t, v_t = torch.linalg.svd(a_t, full_matrices=False)
+        u_t, s_t, v_t = torch.linalg.svd(a_t.float(), full_matrices=False)
         r = u_s.shape[1]
         s_root = torch.sqrt(s_t[:r].clamp(min=1e-8))
         u_ref = u_t[:, :r] * s_root.unsqueeze(0)
